Checking if blank CNAME works.
Basically, a second ago I tried using cia.okstate.edu in here, because of a fundamental misunderstanding of the function of the CNAME, and now it redirects to cia.okstate.edu, which does not exist in part because it is just meant to redirect to here. So I tried putting osucia.github.io in here instead, but that didn't work to update the redirect at the DNS level. So I'm trying a blank one.